Generic Worker - make artifacts a task feature
