+++ changelog/Whg-6bCnTgSZzeNFwnSdXg.md
@@ -1,0 +1,4 @@
+audience: developers
+level: patch
+---
+Generic Worker now handles artifact upload inside a task feature rather than the main processing loop of the task execution. This refactor improves code modularity.
